Wrapper for Camino dt2nii function

Performs the Camino dt2nii function

camino_dt2nii(infile, header, inputmodel = c("dt", "twotensor",
  "threetensor", "multitensor"), outputroot = NULL,
  outputdatatype = NULL, maxcomponents = NULL, gzip = TRUE,
  verbose = TRUE)

Arguments

infile
Output from camino_modelfit

header
NIfTI image with header information

inputmodel
model used for tensors

outputroot
Output file root. If there is one component, or output is to a vector, the output image is called root.ext, otherwise output is numbered by component.

outputdatatype
Specifies the data type of the output image. The data type can be any of the following strings: "byte", "short", "int", "long", "float" or "double".

maxcomponents
Maximum tensor components, for use with input model "multitensor".

gzip
Set gzip compression for the output image, if supported by the image format.

verbose
print diagnostic messages