Parse arguments from Camino

Wrapper for finding the arguments for the Camino function

parse_camino_args(funcname)

Arguments

funcname
Name of function to be called

Value

Character vector

Examples

read_camino_helpfile("voxel2scanner")
#>  [1] ".\\\" $Id: voxel2scanner.1,v 1.2 2005/06/15 08:42:33 ucacdxa Exp $"                     
#>  [2] ""                                                                                       
#>  [3] ".TH voxel2scanner 1"                                                                    
#>  [4] ""                                                                                       
#>  [5] ".SH NAME"                                                                               
#>  [6] "voxel2scanner \\- Converts voxel-order data to scanner-order data."                     
#>  [7] ""                                                                                       
#>  [8] ".SH SYNOPSIS"                                                                           
#>  [9] ".B voxel2scanner"                                                                       
#> [10] "-voxels <number of voxels> -components <measurements per voxel> [\\fIoptions\\fR]"      
#> [11] ""                                                                                       
#> [12] ".SH DESCRIPTION"                                                                        
#> [13] "Converts voxel-order data files, used by Camino, to scanner order. See scanner2voxel(1)"
#> [14] "for a description of these formats."                                                    
#> [15] ""                                                                                       
#> [16] "The output data type is big endian floating points."                                    
#> [17] ""                                                                                       
#> [18] ".SH EXAMPLES"                                                                           
#> [19] "Convert voxel-order data file VoxelOrder.Bfloat, for which the measurement volume has"  
#> [20] "size 128x128x60 and the acquisition scheme acquires 6 measurements with b=0 and 54"     
#> [21] "diffusion-weighted measurements, to scanner order:"                                     
#> [22] ""                                                                                       
#> [23] "voxel2scanner -voxels 983040 -components 60 < VoxelOrder.Bfloat > ScannerOrder.img"     
#> [24] ""                                                                                       
#> [25] "Equivalently:"                                                                          
#> [26] ""                                                                                       
#> [27] "voxel2scanner -voxels 983040 -components 60 -inputfile VoxelOrder.Bfloat >"             
#> [28] "ScannerOrder.img"                                                                       
#> [29] ""                                                                                       
#> [30] ""                                                                                       
#> [31] "Fit the diffusion tensor to the scanner-order data in ScannerOrder.img, assuming the"   
#> [32] "acquisition scheme is specified in A.scheme, and convert the output to scanner order:"  
#> [33] ""                                                                                       
#> [34] "scanner2voxel -voxels 983040 -components 60 < ScannerOrder.img | dtfit - A.scheme"      
#> [35] "-nonlinear | voxel2scanner -voxels 983040 -components 8 -inputdatatype double >"        
#> [36] "DiffTensorA.img"                                                                        
#> [37] ""                                                                                       
#> [38] ".SH OPTIONS"                                                                            
#> [39] ".TP"                                                                                    
#> [40] ".B \\-voxels <number of voxels>"                                                        
#> [41] "The number of voxels in the input data."                                                
#> [42] ""                                                                                       
#> [43] ".TP"                                                                                    
#> [44] ".B \\-components <number of measurements>"                                              
#> [45] "The number of measurements per voxel."                                                  
#> [46] ""                                                                                       
#> [47] ".TP"                                                                                    
#> [48] ".B \\-inputfile <\\fIinput filename\\fR>"                                               
#> [49] "The name of the input file."                                                            
#> [50] ""                                                                                       
#> [51] ".TP"                                                                                    
#> [52] ".B \\-inputdatatype <\\fIdata type of input\\fR>"                                       
#> [53] "Specifies the data type of the input file.  The data type can be any of the following"  
#> [54] "strings: \"char\", \"short\", \"int\", \"long\", \"float\" or \"double\"."              
#> [55] ""                                                                                       
#> [56] ".TP"                                                                                    
#> [57] ".B \\-outputdatatype <\\fIdata type of output\\fR>"                                     
#> [58] "Specifies the data type of the output data.  The data type can be any of the following" 
#> [59] "strings: \"char\", \"short\", \"int\", \"long\", \"float\" or \"double\"."              
#> [60] ""                                                                                       
#> [61] ".SH \"AUTHORS\""                                                                        
#> [62] "Daniel Alexander <camino@cs.ucl.ac.uk>"                                                 
#> [63] ""                                                                                       
#> [64] ".SH \"SEE ALSO\""                                                                       
#> [65] "scanner2voxel(1)"                                                                       
#> [66] ""                                                                                       
#> [67] ".SH BUGS"